iOS + Python test
Started the iOS project, found a nice TCP/UDP library (we are using TCP
as it’s important that we receive all of our data). Also tested the
nature of TCP - the protocol can split the data it receives into as
many chunks as it wants. Just means we’ll have to have a “start” and
“end” designator for each poll.